refactor: 이동 전략의 패키지 및 파일명 변경
