+++ src/main/java/domain/move/strategy/Movement.java
@@ -1,11 +1,11 @@
-package domain.move.movement;
+package domain.move.strategy;
 
 import domain.board.Intersection;
 import domain.game.Side;
 import domain.move.Path;
 import java.util.List;
 
-public abstract class PieceMovement {
+public abstract class Movement {
 
     public final List<Path> movablePaths(Intersection from, Side side) {
         List<Path> allPaths = candidatePaths(from, side);
