+++ src/main/java/domain/move/strategy/SingleStepMovement.java
@@ -1,4 +1,4 @@
-package domain.move.movement;
+package domain.move.strategy;
 
 import domain.board.Intersection;
 import domain.direction.Direction;
@@ -9,7 +9,7 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
-public final class SingleStepMovement extends PieceMovement {
+public final class SingleStepMovement extends Movement {
 
     private static final MoveAmount MOVE_AMOUNT = new MoveAmount(1);
