+++ src/test/java/domain/move/strategy/OrthogonalThenTwoDiagonalMovementTest.java
@@ -1,13 +1,11 @@
-package domain.move.movement;
+package domain.move.strategy;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import domain.board.Intersection;
 import domain.game.Side;
 import domain.move.Path;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.stream.IntStream;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -17,26 +15,25 @@ import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.EnumSource;
 import org.junit.jupiter.params.provider.MethodSource;
 
-@DisplayName("직선 이동 테스트")
-class StraightLineMovementTest {
+@DisplayName("밭전자 이동 테스트")
+class OrthogonalThenTwoDiagonalMovementTest {
 
-    private StraightLineMovement movement;
+    private Movement movement;
 
     @BeforeEach
     void setUp() {
-        movement = new StraightLineMovement();
+        movement = new OrthogonalThenTwoDiagonalMovement();
     }
 
-    @DisplayName("앞, 뒤, 양 옆으로 칸 수 제약없이 이동이 가능하다")
+    @DisplayName("앞, 뒤, 양 옆 방향마다 두 갈래의 밭전 방향으로 이동이 가능하다")
     @Nested
-    class 모든_빙향으로_칸_수_제약없이_이동이_가능하다 {
+    class 모든_방향마다_두_갈래의_밭전_방향으로_이동이_가능하다 {
 
         @DisplayName("진영에 상관없이 동일하게 적용된다")
         @ParameterizedTest
         @EnumSource(Side.class)
         void 진영에_상관없이_동일하게_적용된다(Side side) {
             Intersection currentIntersection = new Intersection(5, 5);
-            List<Intersection> expected = allDirectionIntersectionsExcludeCurrentIntersection(currentIntersection);
 
             List<Path> movablePaths = movement.movablePaths(currentIntersection, side);
             List<Intersection> movableDestinations = movablePaths.stream()
@@ -44,7 +41,16 @@ class StraightLineMovementTest {
                     .toList();
 
             assertThat(movableDestinations)
-                    .containsExactlyInAnyOrderElementsOf(expected);
+                    .containsExactlyInAnyOrder(
+                            new Intersection(2, 3),
+                            new Intersection(2, 7),
+                            new Intersection(8, 3),
+                            new Intersection(8, 7),
+                            new Intersection(3, 2),
+                            new Intersection(3, 8),
+                            new Intersection(7, 2),
+                            new Intersection(7, 8)
+                    );
         }
     }
 
@@ -74,35 +80,14 @@ class StraightLineMovementTest {
         private static Stream<Arguments> allBorderIntersections() {
             return Stream.of(
                     Arguments.of(new Intersection(1, 1),
-                            allDirectionIntersectionsExcludeCurrentIntersection(new Intersection(1, 1))),
+                            List.of(new Intersection(3, 4), new Intersection(4, 3))),
                     Arguments.of(new Intersection(1, 9),
-                            allDirectionIntersectionsExcludeCurrentIntersection(new Intersection(1, 9))),
+                            List.of(new Intersection(3, 6), new Intersection(4, 7))),
                     Arguments.of(new Intersection(10, 1),
-                            allDirectionIntersectionsExcludeCurrentIntersection(new Intersection(10, 1))),
+                            List.of(new Intersection(7, 3), new Intersection(8, 4))),
                     Arguments.of(new Intersection(10, 9),
-                            allDirectionIntersectionsExcludeCurrentIntersection(new Intersection(10, 9)))
+                            List.of(new Intersection(8, 6), new Intersection(7, 7)))
             );
         }
-    }
-
-    private static List<Intersection> allDirectionIntersectionsExcludeCurrentIntersection(
-            Intersection currentIntersection
-    ) {
-        int currentRow = currentIntersection.row();
-        int currentFile = currentIntersection.file();
-
-        List<Intersection> expected = new ArrayList<>();
-
-        IntStream.rangeClosed(1, 10)
-                .filter(row -> row != currentRow)
-                .mapToObj(row -> new Intersection(row, currentFile))
-                .forEach(expected::add);
-
-        IntStream.rangeClosed(1, 9)
-                .filter(file -> file != currentFile)
-                .mapToObj(file -> new Intersection(currentRow, file))
-                .forEach(expected::add);
-
-        return expected;
     }
 }
