+++ src/test/java/domain/move/strategy/StraightLineMovementTest.java
@@ -1,11 +1,13 @@
-package domain.move.movement;
+package domain.move.strategy;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import domain.board.Intersection;
 import domain.game.Side;
 import domain.move.Path;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.IntStream;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -15,25 +17,26 @@ import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.EnumSource;
 import org.junit.jupiter.params.provider.MethodSource;
 
-@DisplayName("모든 방향 1칸 이동 테스트")
-class SingleStepMovementTest {
+@DisplayName("직선 이동 테스트")
+class StraightLineMovementTest {
 
-    private SingleStepMovement movement;
+    private StraightLineMovement movement;
 
     @BeforeEach
     void setUp() {
-        movement = new SingleStepMovement();
+        movement = new StraightLineMovement();
     }
 
-    @DisplayName("앞, 뒤, 양 옆으로 1칸 이동이 가능하다")
+    @DisplayName("앞, 뒤, 양 옆으로 칸 수 제약없이 이동이 가능하다")
     @Nested
-    class 모든_빙향으로_1칸_이동이_가능하다 {
+    class 모든_빙향으로_칸_수_제약없이_이동이_가능하다 {
 
         @DisplayName("진영에 상관없이 동일하게 적용된다")
         @ParameterizedTest
         @EnumSource(Side.class)
         void 진영에_상관없이_동일하게_적용된다(Side side) {
             Intersection currentIntersection = new Intersection(5, 5);
+            List<Intersection> expected = allDirectionIntersectionsExcludeCurrentIntersection(currentIntersection);
 
             List<Path> movablePaths = movement.movablePaths(currentIntersection, side);
             List<Intersection> movableDestinations = movablePaths.stream()
@@ -41,12 +44,7 @@ class SingleStepMovementTest {
                     .toList();
 
             assertThat(movableDestinations)
-                    .containsExactlyInAnyOrder(
-                            new Intersection(4, 5),
-                            new Intersection(6, 5),
-                            new Intersection(5, 4),
-                            new Intersection(5, 6)
-                    );
+                    .containsExactlyInAnyOrderElementsOf(expected);
         }
     }
 
@@ -76,14 +74,35 @@ class SingleStepMovementTest {
         private static Stream<Arguments> allBorderIntersections() {
             return Stream.of(
                     Arguments.of(new Intersection(1, 1),
-                            List.of(new Intersection(1, 2), new Intersection(2, 1))),
+                            allDirectionIntersectionsExcludeCurrentIntersection(new Intersection(1, 1))),
                     Arguments.of(new Intersection(1, 9),
-                            List.of(new Intersection(1, 8), new Intersection(2, 9))),
+                            allDirectionIntersectionsExcludeCurrentIntersection(new Intersection(1, 9))),
                     Arguments.of(new Intersection(10, 1),
-                            List.of(new Intersection(9, 1), new Intersection(10, 2))),
+                            allDirectionIntersectionsExcludeCurrentIntersection(new Intersection(10, 1))),
                     Arguments.of(new Intersection(10, 9),
-                            List.of(new Intersection(9, 9), new Intersection(10, 8)))
+                            allDirectionIntersectionsExcludeCurrentIntersection(new Intersection(10, 9)))
             );
         }
+    }
+
+    private static List<Intersection> allDirectionIntersectionsExcludeCurrentIntersection(
+            Intersection currentIntersection
+    ) {
+        int currentRow = currentIntersection.row();
+        int currentFile = currentIntersection.file();
+
+        List<Intersection> expected = new ArrayList<>();
+
+        IntStream.rangeClosed(1, 10)
+                .filter(row -> row != currentRow)
+                .mapToObj(row -> new Intersection(row, currentFile))
+                .forEach(expected::add);
+
+        IntStream.rangeClosed(1, 9)
+                .filter(file -> file != currentFile)
+                .mapToObj(file -> new Intersection(currentRow, file))
+                .forEach(expected::add);
+
+        return expected;
     }
 }
